Preps and documentation for running in the containers.
Running the code in the containers have an advantage of reproducibility and also, if done right way,
improves security. virtualenv environment doesn't protect the data of the developer from being
stolen or tampered with. The Docker/Podman containers are helping with this issue and are quite
good tradeof between ease of use, speed of start and security.

Running new code in the container is good practice for the software developer that don't want
his repositories to get compromised with supply-chain compromise attacks.